+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Avoid recalculations on pause component.
 - Avoid RwLock and Mutex from tokio when possible in Hydrolink.
 - Change how the Lavalink REST client makes requests.
+- Get the VoiceState from the ConnectionInfo instead of storing it.
 - Use Cow<str> instead of String when possible.
 - Use dynfmt crate instead of `String::replace`.
 - Use Hydrogen's green instead of Discord's blue as primary color.